refactor: organize app query modules
