+++ docs/official/architecture.md
@@ -76,7 +76,7 @@
 - **projectId**：URL 中使用 `.spherse/project.yaml` 中的稳定 id（nanoid 8 位 token）作为路由参数，替代旧的基于目录名生成的 projectKey；全链路（core → server → IPC → renderer 路由 → localStorage key）统一使用 projectId 作为唯一身份标识
 - **项目内 lastRoute**：每个打开项目的子页面路由持久化在 renderer localStorage（`spherse:last-route:<projectId>`），不写入 electron-store。应用启动、项目切换和关闭当前项目后的下一个项目导航都会恢复该项目的 lastRoute；关闭项目时清理对应的 localStorage key
 - **应用级 store**：`app-store.ts` 管理打开项目集合、当前项目、restore/open/close/reveal 等 Electron IPC 相关动作，并持久化左侧 side panel 固定/自动收起偏好
-- **服务端状态缓存**：renderer 根节点挂载 TanStack Query；`lib/project-queries.ts` 统一管理按 project 隔离的 agents、sessions 与 session 分页缓存、mutation 后失效、bus/reconnect 刷新和 UI SDK 的命令式缓存访问；`lib/content-queries.ts` 按 project + path 缓存文件内容、目录列表和完整文件路径索引，使主内容页、浮窗、文件树与文件搜索共享请求和快照。`ContentQueryBridge` 将 fs-watch/reconnect 集中转换为 cache invalidation，树的展开/选择及编辑器 draft/conflict 仍保留为本地交互状态。关闭项目时移除对应 project query cache。`project-data-store.ts` 不再复制服务端快照，只保留初始消息、streaming session id 与 trigger 启用标记等前端运行时投影。
+- **服务端状态缓存**：renderer 根节点挂载 TanStack Query；`queries/project.ts` 统一管理按 project 隔离的 agents、sessions 与 session 分页缓存、mutation 后失效、bus/reconnect 刷新和 UI SDK 的命令式缓存访问；`queries/content.ts` 按 project + path 缓存文件内容、目录列表和完整文件路径索引，使主内容页、浮窗、文件树与文件搜索共享请求和快照。`ContentQueryBridge` 将 fs-watch/reconnect 集中转换为 cache invalidation，树的展开/选择及编辑器 draft/conflict 仍保留为本地交互状态。关闭项目时移除对应 project query cache。`project-data-store.ts` 不再复制服务端快照，只保留初始消息、streaming session id 与 trigger 启用标记等前端运行时投影。
 - **项目 UI store**：折叠状态、浮窗会话状态等纯 UI 状态按 projectId 管理，通过 renderer localStorage 持久化。折叠状态在 `features/agent-session-list/store.ts`，浮窗会话状态在 `features/floating-chat/store.ts`，浮窗内容浏览器（多窗口，按 projectId → filePath 映射）在 `features/floating-content-browser/store.ts`。浮窗 chrome（拖拽/调整大小）由共享的 `components/floating-frame/` 提供，floating-chat 与 floating-content-browser 通过 `hookPrefix` 参数复用并各自生成 `data-chat-float-*` / `data-content-float-*` 主题钩子
 - **局部状态边界**：文件编辑 dirty/conflict、弹窗表单等短生命周期状态保留在对应组件或 feature hook 内；Chat 输入框草稿按 sessionId 缓存在 renderer `localStorage`，用于 session 切换和应用重启后的草稿恢复；欢迎页设置与主题设置 dialog 的打开状态保留在 `ActivityBar` 内（通过「设置」二级菜单触发，仅 active 项目可见），设置变更通过 `lib/events.ts` 中的 renderer 自定义事件通知：欢迎页重新读取项目配置、`useCustomTheme` 重新挂载主题 `<link>` 热更新
 - **Chat run / WebSocket 生命周期**：Core `AgentRunner` 只拥有 agent 上下文、执行与消息持久化，通过单次 `sendMessage` callback 输出事件（并发 turn 有 in-flight guard，二次调用抛 ValidationError）；server `ChatSessionHub` 按 project/session 建立 transport channel，独立负责共享 restore、当前 run 状态、事件广播、压缩事件快照和无 attachment 时的空闲释放。hub 实例由 `server/index.ts` 创建并共享给 `ws-chat.ts` 与 sessions 路由：UI SDK 的静默发送（`sendMessage` 带 `open:false` 且目标会话未 attach WS 时）走 `POST /api/projects/:projectId/agents/:agentId/sessions/:id/messages`，经 hub 的 `startDetachedRun`（fire-and-ack：restore ready + running 检查后启动 run 即返回，run 失败仅 log + 向 subscriber 发 error 事件）在同一 channel 上执行，与已打开的 WS 共享 run 序列化并 fan-out 事件；channel running 时返回 409。WebSocket 是 channel 的可替换 attachment，socket close 只解除 attachment；若 agent 正在运行，Core run 会继续执行并持久化消息，重连后由 hub 先重放当前 run 快照，再发送 `run_status` 对账运行状态。WebSocket 连接具备心跳保活与指数退避重连（1/2/5/10/30s）：每 30s 发 ping，只有连续 60s 等不到 pong 才关闭；renderer suspend 导致 timer 大幅跳跃时会重置探测窗口，避免恢复后误判超时。socket 发送失败被隔离在 server attachment 边界，Core 则保证 `message_end` 先持久化再回调，连接故障不能打断 agent 或阻止最终消息落库。
