+++ packages/app/src/main.tsx
@@ -4,7 +4,7 @@ import { QueryClientProvider } from "@tanstack/react-query";
 import type { HostBridge } from "./lib/host-bridge";
 import { HostBridgeProvider } from "./context/host-bridge-context";
 import { router } from "./router";
-import { queryClient } from "./lib/query-client";
+import { queryClient } from "./queries/client";
 import "./styles.css";
 
 export function createAppRoot(bridge: HostBridge): Root {
