+++ packages/app/src/ui-sdk/handlers/float-session.ts
@@ -4,7 +4,7 @@ import { toast } from "sonner";
 import { translate, normalizeLocale } from "@spherse/i18n";
 import { useSettingsStore } from "../../stores/settings-store";
 import { openChat } from "./open-chat";
-import { ensureProjectSession } from "../../lib/project-queries";
+import { ensureProjectSession } from "../../queries/project";
 
 registerAction("floatSession", async (params, ctx) => {
   const { sessionId } = params as { sessionId: string };
