+++ packages/app/src/ui-sdk/handlers/open-session.ts
@@ -4,7 +4,7 @@ import { toast } from "sonner";
 import { translate, normalizeLocale } from "@spherse/i18n";
 import { useSettingsStore } from "../../stores/settings-store";
 import { openChat } from "./open-chat";
-import { ensureProjectSession } from "../../lib/project-queries";
+import { ensureProjectSession } from "../../queries/project";
 
 registerAction("openSession", async (params, ctx) => {
   const { sessionId, float } = params as { sessionId: string; float?: boolean };
